+++ app/layout.js
@@ -1,0 +1,13 @@
+import './styles/globals.css'
+
+export default function RootLayout({
+    // Layouts must accept a children prop.
+    // This will be populated with nested layouts or pages
+    children,
+}) {
+    return (
+        <html lang="en">
+            <body>{children}</body>
+        </html>
+    )
+}
