+++ package.json
@@ -12,8 +12,8 @@
     "eventsource-parser": "^1.1.1",
     "next": "^14.0.3",
     "openai": "^4.19.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-particles": "^2.10.1",
     "tsparticles": "^2.10.1",
     "tsparticles-engine": "^2.10.0"
