+++ tsconfig.json
@@ -17,13 +17,19 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "jsx": "preserve",
-    "incremental": true
+    "incremental": true,
+    "plugins": [
+      {
+        "name": "next"
+      }
+    ]
   },
   "include": [
     "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx",
-    "pages/completions.js"
+    "app/completions/page.js",
+    ".next/types/**/*.ts"
   ],
   "exclude": [
     "node_modules"
